Create SimpleResponsiveLayout.html
this is just an example how the frontend of my Mitfahr-App will look like. the Gray-Part is where you log yourself in .The link is for the registration
the white part is where the map is supposed to be set and the orange part is for the result of the researches.
This is just like a simulation and i made it responsive .
Stay tuned guys i'm coming back soon with the end of the frontend of my app it will be very nice.
I'm actually using HTML, CSS , Javascript ,Docker and MySQL for the Frontend and backend of my App.